feat(web): localize week views and introduce entry values
Adds localization support to the week grid and mission summaries, improving clarity for non-French users. Summarizes tracked entries using API-provided `value` figures, ensuring consistency with backend pricing logic. Enhances UX with clearer displays and adjusted contrast for pending entries.

+++ apps/api/app/Domain/Invoices/Actions/ValueTrackedTime.php
@@ -95,7 +95,8 @@ class ValueTrackedTime
 
         return [
             'value' => $rate->multiply($numerator)->divide($denominator, MoneyPhp::ROUND_HALF_UP),
-            'days' => $numerator / $denominator,
+            // PHP's / hands back an int when the division comes out even.
+            'days' => (float) ($numerator / $denominator),
         ];
     }
